# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NarrativeRequest": {
        "properties": {
          "compareGroup": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparegroup"
          },
          "keys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keys"
          },
          "regionName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regionname"
          },
          "university": {
            "title": "University",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          },
          "years": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "추이·평균·순위에 남길 연도. 생략하면 전 연도. **기준 연도(year)가 반드시 포함되어야 한다.**",
            "title": "Years"
          }
        },
        "required": [
          "university",
          "year"
        ],
        "title": "NarrativeRequest",
        "type": "object"
      },
      "NarrativeResponse": {
        "description": "생성된 서술.\n\n키가 4종 전부 오지 않을 수 있다 — 일부만 요청했거나, 일부만 실패했을 때다.\n`failed` 에 실패한 키와 이유를 담는다. 조용히 빈 문자열을 돌려주면\n사용자는 GPT 가 \"아무 말도 하지 않았다\" 고 오해한다.",
        "properties": {
          "failed": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Failed",
            "type": "object"
          },
          "narratives": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Narratives",
            "type": "object"
          }
        },
        "required": [
          "narratives",
          "failed"
        ],
        "title": "NarrativeResponse",
        "type": "object"
      },
...
      "ReportRequest": {
        "properties": {
          "compareGroup": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparegroup"
          },
          "narratives": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Narratives",
            "type": "object"
          },
          "regionName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regionname"
          },
          "university": {
            "title": "University",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          },
          "years": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "추이·평균·순위에 남길 연도. 생략하면 전 연도. **기준 연도(year)가 반드시 포함되어야 한다.**",
            "title": "Years"
          }
        },
        "required": [
          "university",
          "year"
        ],
        "title": "ReportRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "대학알리미 전임교원 연구실적 분석. core/ 의 순수 함수를 감싼다.",
    "title": "연구실적 분석 포털 API",
    "version": "6.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chart/{kind}.png": {
      "get": {
        "description": "Word 보고서에 들어가는 것과 **같은** PNG 를 돌려준다.\n\n화면 차트를 recharts 로 따로 그리면 사용자가 본 그림과 문서에 실리는\n그림이 갈라진다. Streamlit 판은 같은 PNG 를 화면과 문서가 공유했고\n(\"5종 차트를 확인하세요. 보고서에 그대로 삽입됩니다\"), 그 계약을 지킨다.\n\n추이 차트만은 화면에서 인터랙티브(recharts)로도 보여준다 — 값을 짚어\n읽는 용도다. 나머지 4종은 이 PNG 가 유일한 표현이다.",
        "operationId": "get_chart_api_chart__kind__png_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "university",
            "required": true,
            "schema": {
              "title": "University",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "description": "비교군 대학 이름. **반드시 보고서와 같은 값을 보내야 한다** — 생략하면 서버 기본 비교군으로 그려져 Word 와 그림이 달라진다.",
            "in": "query",
            "name": "compareGroup",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "비교군 대학 이름. **반드시 보고서와 같은 값을 보내야 한다** — 생략하면 서버 기본 비교군으로 그려져 Word 와 그림이 달라진다.",
              "title": "Comparegroup"
            }
          },
          {
            "description": "분석 연도. 비교군과 같은 이유로 **보고서와 같은 값을 보내야 한다** — 생략하면 전 연도로 그려져 화면의 추이 차트와 Word 가 갈라진다.",
            "in": "query",
            "name": "years",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "분석 연도. 비교군과 같은 이유로 **보고서와 같은 값을 보내야 한다** — 생략하면 전 연도로 그려져 화면의 추이 차트와 Word 가 갈라진다.",
              "title": "Years"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chart",
        "tags": [
          "report"
        ]
      }
    },
...
    "/api/narrative": {
      "post": {
        "operationId": "post_narrative_api_narrative_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Narrative",
        "tags": [
          "report"
        ]
      }
    },
...
    "/api/report": {
      "post": {
        "operationId": "post_report_api_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Report",
        "tags": [
          "report"
        ]
      }
    },
...
  }
}